MR-45-AddCategoryUI updating add category view with the custom functions and use the dimens file
