+++ app/src/main/java/com/example/myriyal/screens/categories/presentation/components/CustomDropdown.kt
@@ -1,14 +1,16 @@
 package com.example.myriyal.screens.categories.presentation.components
+
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.heightIn
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowDropDown
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -16,7 +18,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
-import com.example.myriyal.R
+import androidx.compose.ui.unit.dp
 import com.example.myriyal.screenComponent.CustomTextField
 
 @Composable
@@ -47,10 +49,9 @@ fun <T> CustomDropdown(
             expanded = expanded,
             onDismissRequest = { expanded = false }
         ) {
-           list.forEach { Item/*type*/ ->
-
+           list.forEach { Item->
                 DropdownMenuItem(
-                    text = { Item/*type.name*/ },
+                    text = { Item },
                     onClick = {
                         onSelect(Item)
                         expanded = false
@@ -58,5 +59,25 @@ fun <T> CustomDropdown(
                 )
             }
         }
+
+        // DropdownMenu updated version to set the max hit of the list but it's lead to a crash!
+//        DropdownMenu(
+//            expanded = expanded,
+//            onDismissRequest = { expanded = false }
+//        ) {
+//            LazyColumn(
+//                modifier = Modifier.heightIn(max = 200.dp) // adjust max height as needed
+//            ) {
+//                items(list) { item ->
+//                    DropdownMenuItem(
+//                        text = { item }, // assuming item has a `name` property
+//                        onClick = {
+//                            onSelect(item)
+//                            expanded = false
+//                        }
+//                    )
+//                }
+//            }
+//        }
     }
 }
