+++ app/src/main/res/values/dimens.xml
@@ -42,7 +42,7 @@
     <integer name="columnPadding">32</integer>
     <integer name="verticalSpacer">16</integer>
     <integer name="RoundedCornerShape">12</integer>
-    <integer name="colorFieldWidth">280</integer>
+    <integer name="colorFieldWidth">380</integer>
     <integer name="colorFieldHeight">55</integer>
     <integer name="colorSampleSize">32</integer>
     <integer name="tonalElevation">8</integer>
